First pass implementing the notification router.
Turns out - Fabric's notification router code I copied is out of date, and none of the text matches up with the callback text.  Soooo...

+++ src/MaxConversionFns.cpp
@@ -241,16 +241,16 @@ void ConvertToRTVal(int param, FabricCore::RTVal& rtVal)
 		rtVal = FabricCore::RTVal::ConstructUInt32(GetClient(), param);
 
 	else if (strcmp(spliceType, "SInt8") == 0)
-		rtVal = FabricCore::RTVal::ConstructSInt8(GetClient(), param);
+		rtVal = FabricCore::RTVal::ConstructSInt8(GetClient(), (int8_t)param);
 	else if (strcmp(spliceType, "SInt16") == 0)
-		rtVal = FabricCore::RTVal::ConstructSInt16(GetClient(), param);
+		rtVal = FabricCore::RTVal::ConstructSInt16(GetClient(), (int16_t)param);
 	else if (strcmp(spliceType, "SInt64") == 0)
 		rtVal = FabricCore::RTVal::ConstructSInt64(GetClient(), param);
 
 	else if (strcmp(spliceType, "UInt8") == 0)
-		rtVal = FabricCore::RTVal::ConstructUInt8(GetClient(), param);
+		rtVal = FabricCore::RTVal::ConstructUInt8(GetClient(), (uint8_t)param);
 	else if (strcmp(spliceType, "UInt16") == 0)
-		rtVal = FabricCore::RTVal::ConstructUInt16(GetClient(), param);
+		rtVal = FabricCore::RTVal::ConstructUInt16(GetClient(), (uint16_t)param);
 	else if (strcmp(spliceType, "UInt64") == 0)
 		rtVal = FabricCore::RTVal::ConstructUInt64(GetClient(), param);
 
@@ -685,20 +685,20 @@ void SpliceToMaxValue(const FabricCore::RTVal& rtVal, int& param)
 	else if (strcmp(spliceType, "SInt16") == 0)
 		param = ncrtVal.getSInt16();
 	else if (strcmp(spliceType, "SInt64") == 0)
-		param = ncrtVal.getSInt64();
+		param = (int)ncrtVal.getSInt64();
 
 	else if (strcmp(spliceType, "UInt8") == 0)
-		param = ncrtVal.getUInt8();
+		param = (int)ncrtVal.getUInt8();
 	else if (strcmp(spliceType, "UInt16") == 0)
-		param = ncrtVal.getUInt16();
+		param = (int)ncrtVal.getUInt16();
 	else if (strcmp(spliceType, "UInt64") == 0)
-		param = ncrtVal.getUInt64();
+		param = (int)ncrtVal.getUInt64();
 
 	// These last two param types may be deprecated
 	else if (strcmp(spliceType, "Size") == 0)
-		param = ncrtVal.getUInt64();
+		param = (int)ncrtVal.getUInt64();
 	else if (strcmp(spliceType, "Index") == 0)
-		param = ncrtVal.getUInt64();
+		param = (int)ncrtVal.getUInt64();
 
 	//param = const_cast<FabricCore::RTVal&>(rtVal).getSInt32();
 }
@@ -711,7 +711,7 @@ void SpliceToMaxValue(const FabricCore::RTVal& rtVal, float& param)
 	if (strcmp(spliceType, "Float32") == 0)
 		param = ncrtVal.getFloat32();
 	if (strcmp(spliceType, "Float64") == 0)
-		param = ncrtVal.getFloat64();
+		param = (float)ncrtVal.getFloat64();
 
 	//param = static_cast<float>(const_cast<FabricCore::RTVal&>(rtVal).getFloat32());
 }
